Fixed bugs in compare_conditions, synapto_display, and added script
Fixed bug in compare conditions.m script.
-The Overlap_with_Blue filter was in a nested if statement that couldn't be reached if the Overlap_with_Green filter was turned on.
- Added columns to record the number of localisations for each synaptosome region of interest.
Added functionality in synapto-display script to only generate images of the regions which have been manually.
Added brief script to re-run the plotting part of the compare_conditions.m script from a .csv file of results.
Modified generateImage.m plot to create images with the intensity based on the number of localisations and not the intensity of the localisations.