Added rejection rate for compare_conditions
Added lines in compareConditions.m to calculate the rejection rate (% of localisations thrown away) in each "detected" region of interest for each channel in all conditions. This can help us understand how much data we're NOT using.

Also added a few lines in Synapto-analysis to make it easier to change the localisation fit filters to find out which ones are the most sensitive parameters.